requirements.txt
sanic
motor
python-dotenv
pydantic